+++ bdk-ffi/src/bitcoin.rs
@@ -1,27 +1,33 @@
 use crate::error::{
-    AddressParseError, FeeRateError, FromScriptError, PsbtError, PsbtParseError, TransactionError,
+    AddressParseError, ExtractTxError, FeeRateError, FromScriptError, HashParseError, PsbtError,
+    PsbtParseError, TransactionError,
 };
 use crate::error::{ParseAmountError, PsbtFinalizeError};
-use crate::{impl_from_core_type, impl_into_core_type};
+use crate::{impl_from_core_type, impl_hash_like, impl_into_core_type};
 
 use bdk_wallet::bitcoin::address::NetworkChecked;
 use bdk_wallet::bitcoin::address::NetworkUnchecked;
 use bdk_wallet::bitcoin::address::{Address as BdkAddress, AddressData as BdkAddressData};
 use bdk_wallet::bitcoin::blockdata::block::Header as BdkHeader;
+use bdk_wallet::bitcoin::consensus::encode::deserialize;
 use bdk_wallet::bitcoin::consensus::encode::serialize;
 use bdk_wallet::bitcoin::consensus::Decodable;
+use bdk_wallet::bitcoin::hashes::sha256::Hash as BitcoinSha256Hash;
+use bdk_wallet::bitcoin::hashes::sha256d::Hash as BitcoinDoubleSha256Hash;
 use bdk_wallet::bitcoin::io::Cursor;
-use bdk_wallet::bitcoin::psbt::ExtractTxError;
 use bdk_wallet::bitcoin::secp256k1::Secp256k1;
 use bdk_wallet::bitcoin::Amount as BdkAmount;
+use bdk_wallet::bitcoin::BlockHash as BitcoinBlockHash;
 use bdk_wallet::bitcoin::FeeRate as BdkFeeRate;
 use bdk_wallet::bitcoin::Network;
+use bdk_wallet::bitcoin::OutPoint as BdkOutPoint;
 use bdk_wallet::bitcoin::Psbt as BdkPsbt;
 use bdk_wallet::bitcoin::ScriptBuf as BdkScriptBuf;
 use bdk_wallet::bitcoin::Transaction as BdkTransaction;
 use bdk_wallet::bitcoin::TxIn as BdkTxIn;
 use bdk_wallet::bitcoin::TxOut as BdkTxOut;
-use bdk_wallet::bitcoin::{OutPoint as BdkOutPoint, Txid};
+use bdk_wallet::bitcoin::Txid as BitcoinTxid;
+use bdk_wallet::bitcoin::Wtxid as BitcoinWtxid;
 use bdk_wallet::miniscript::psbt::PsbtExt;
 use bdk_wallet::serde_json;
 
@@ -30,9 +36,12 @@ use std::ops::Deref;
 use std::str::FromStr;
 use std::sync::{Arc, Mutex};
 
-#[derive(Debug, Clone, Eq, PartialEq)]
+/// A reference to an unspent output by TXID and output index.
+#[derive(Debug, Clone, Eq, PartialEq, uniffi:: Record)]
 pub struct OutPoint {
+    /// The transaction.
     pub txid: String,
+    /// The index of the output in the transaction.
     pub vout: u32,
 }
 
@@ -48,7 +57,7 @@ impl From<&BdkOutPoint> for OutPoint {
 impl From<OutPoint> for BdkOutPoint {
     fn from(outpoint: OutPoint) -> Self {
         BdkOutPoint {
-            txid: Txid::from_str(&outpoint.txid).unwrap(),
+            txid: BitcoinTxid::from_str(&outpoint.txid).unwrap(),
             vout: outpoint.vout,
         }
     }
@@ -130,15 +139,20 @@ impl Amount {
 impl_from_core_type!(BdkAmount, Amount);
 impl_into_core_type!(Amount, BdkAmount);
 
-#[derive(Clone, Debug)]
+/// A bitcoin script: https://en.bitcoin.it/wiki/Script
+#[derive(Clone, Debug, uniffi::Object)]
 pub struct Script(pub BdkScriptBuf);
 
+#[uniffi::export]
 impl Script {
+    /// Interpret an array of bytes as a bitcoin script.
+    #[uniffi::constructor]
     pub fn new(raw_output_script: Vec<u8>) -> Self {
         let script: BdkScriptBuf = raw_output_script.into();
         Script(script)
     }
 
+    /// Convert a script into an array of bytes.
     pub fn to_bytes(&self) -> Vec<u8> {
         self.0.to_bytes()
     }
@@ -147,12 +161,22 @@ impl Script {
 impl_from_core_type!(BdkScriptBuf, Script);
 impl_into_core_type!(Script, BdkScriptBuf);
 
+/// Bitcoin block header.
+/// Contains all the block’s information except the actual transactions, but including a root of a merkle tree
+/// committing to all transactions in the block.
+#[derive(uniffi::Record)]
 pub struct Header {
+    /// Block version, now repurposed for soft fork signalling.
     pub version: i32,
+    /// Reference to the previous block in the chain.
     pub prev_blockhash: String,
+    /// The root hash of the merkle tree of transactions in the block.
     pub merkle_root: String,
+    /// The timestamp of the block, as claimed by the miner.
     pub time: u32,
+    /// The target value below which the blockhash must lie.
     pub bits: u32,
+    /// The nonce, selected to obtain a low enough blockhash.
     pub nonce: u32,
 }
 
@@ -169,23 +193,35 @@ impl From<BdkHeader> for Header {
     }
 }
 
-#[derive(Debug)]
+/// The type of address.
+#[derive(Debug, uniffi::Enum)]
 pub enum AddressData {
+    /// Legacy.
     P2pkh { pubkey_hash: String },
+    /// Wrapped Segwit
     P2sh { script_hash: String },
+    /// Segwit
     Segwit { witness_program: WitnessProgram },
 }
 
-#[derive(Debug)]
+/// The version and program of a Segwit address.
+#[derive(Debug, uniffi::Record)]
 pub struct WitnessProgram {
+    /// Version. For example 1 for Taproot.
     pub version: u8,
+    /// The witness program.
     pub program: Vec<u8>,
 }
 
-#[derive(Debug, PartialEq, Eq)]
+/// A bitcoin address
+#[derive(Debug, PartialEq, Eq, uniffi::Object)]
+#[uniffi::export(Eq, Display)]
 pub struct Address(BdkAddress<NetworkChecked>);
 
+#[uniffi::export]
 impl Address {
+    /// Parse a string as an address for the given network.
+    #[uniffi::constructor]
     pub fn new(address: String, network: Network) -> Result<Self, AddressParseError> {
         let parsed_address = address.parse::<bdk_wallet::bitcoin::Address<NetworkUnchecked>>()?;
         let network_checked_address = parsed_address.require_network(network)?;
@@ -193,20 +229,25 @@ impl Address {
         Ok(Address(network_checked_address))
     }
 
+    /// Parse a script as an address for the given network
+    #[uniffi::constructor]
     pub fn from_script(script: Arc<Script>, network: Network) -> Result<Self, FromScriptError> {
         let address = BdkAddress::from_script(&script.0.clone(), network)?;
 
         Ok(Address(address))
     }
 
+    /// Return the `scriptPubKey` underlying an address.
     pub fn script_pubkey(&self) -> Arc<Script> {
         Arc::new(Script(self.0.script_pubkey()))
     }
 
+    /// Return a BIP-21 URI string for this address.
     pub fn to_qr_uri(&self) -> String {
         self.0.to_qr_uri()
     }
 
+    /// Is the address valid for the provided network
     pub fn is_valid_for_network(&self, network: Network) -> bool {
         let address_str = self.0.to_string();
         if let Ok(unchecked_address) = address_str.parse::<BdkAddress<NetworkUnchecked>>() {
@@ -216,6 +257,7 @@ impl Address {
         }
     }
 
+    /// Return the data for the address.
     pub fn to_address_data(&self) -> AddressData {
         match self.0.to_address_data() {
             BdkAddressData::P2pkh { pubkey_hash } => AddressData::P2pkh {
@@ -245,60 +287,124 @@ impl Display for Address {
 impl_from_core_type!(BdkAddress, Address);
 impl_into_core_type!(Address, BdkAddress);
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+/// Bitcoin transaction.
+/// An authenticated movement of coins.
+#[derive(Debug, Clone, PartialEq, Eq, uniffi::Object)]
 pub struct Transaction(BdkTransaction);
 
+#[uniffi::export]
 impl Transaction {
+    /// Creates a new `Transaction` instance from serialized transaction bytes.
+    #[uniffi::constructor]
     pub fn new(transaction_bytes: Vec<u8>) -> Result<Self, TransactionError> {
         let mut decoder = Cursor::new(transaction_bytes);
         let tx: BdkTransaction = BdkTransaction::consensus_decode(&mut decoder)?;
         Ok(Transaction(tx))
     }
 
+    /// Computes the Txid.
+    /// Hashes the transaction excluding the segwit data (i.e. the marker, flag bytes, and the witness fields themselves).
     pub fn compute_txid(&self) -> String {
         self.0.compute_txid().to_string()
     }
 
+    /// Returns the weight of this transaction, as defined by BIP-141.
+    ///
+    /// > Transaction weight is defined as Base transaction size * 3 + Total transaction size (ie.
+    /// > the same method as calculating Block weight from Base size and Total size).
+    ///
+    /// For transactions with an empty witness, this is simply the consensus-serialized size times
+    /// four. For transactions with a witness, this is the non-witness consensus-serialized size
+    /// multiplied by three plus the with-witness consensus-serialized size.
+    ///
+    /// For transactions with no inputs, this function will return a value 2 less than the actual
+    /// weight of the serialized transaction. The reason is that zero-input transactions, post-segwit,
+    /// cannot be unambiguously serialized; we make a choice that adds two extra bytes. For more
+    /// details see [BIP 141](https://github.com/bitcoin/bips/blob/master/bip-0141.mediawiki)
+    /// which uses a "input count" of `0x00` as a `marker` for a Segwit-encoded transaction.
+    ///
+    /// If you need to use 0-input transactions, we strongly recommend you do so using the PSBT
+    /// API. The unsigned transaction encoded within PSBT is always a non-segwit transaction
+    /// and can therefore avoid this ambiguity.
+    #[inline]
     pub fn weight(&self) -> u64 {
         self.0.weight().to_wu()
     }
 
+    /// Returns the total transaction size
+    ///
+    /// Total transaction size is the transaction size in bytes serialized as described in BIP144,
+    /// including base data and witness data.
     pub fn total_size(&self) -> u64 {
         self.0.total_size() as u64
     }
 
+    /// Returns the "virtual size" (vsize) of this transaction.
+    ///
+    /// Will be `ceil(weight / 4.0)`. Note this implements the virtual size as per [`BIP141`], which
+    /// is different to what is implemented in Bitcoin Core.
+    /// > Virtual transaction size is defined as Transaction weight / 4 (rounded up to the next integer).
+    ///
+    /// [`BIP141`]: https://github.com/bitcoin/bips/blob/master/bip-0141.mediawiki
+    #[inline]
     pub fn vsize(&self) -> u64 {
         self.0.vsize() as u64
     }
 
+    /// Checks if this is a coinbase transaction.
+    /// The first transaction in the block distributes the mining reward and is called the coinbase transaction.
+    /// It is impossible to check if the transaction is first in the block, so this function checks the structure
+    /// of the transaction instead - the previous output must be all-zeros (creates satoshis “out of thin air”).
     pub fn is_coinbase(&self) -> bool {
         self.0.is_coinbase()
     }
 
+    /// Returns `true` if the transaction itself opted in to be BIP-125-replaceable (RBF).
+    ///
+    /// # Warning
+    ///
+    /// **Incorrectly relying on RBF may lead to monetary loss!**
+    ///
+    /// This **does not** cover the case where a transaction becomes replaceable due to ancestors
+    /// being RBF. Please note that transactions **may be replaced** even if they **do not** include
+    /// the RBF signal: <https://bitcoinops.org/en/newsletters/2022/10/19/#transaction-replacement-option>.
     pub fn is_explicitly_rbf(&self) -> bool {
         self.0.is_explicitly_rbf()
     }
 
+    /// Returns `true` if this transactions nLockTime is enabled ([BIP-65]).
+    ///
+    /// [BIP-65]: https://github.com/bitcoin/bips/blob/master/bip-0065.mediawiki
     pub fn is_lock_time_enabled(&self) -> bool {
         self.0.is_lock_time_enabled()
     }
 
+    /// The protocol version, is currently expected to be 1 or 2 (BIP 68).
     pub fn version(&self) -> i32 {
         self.0.version.0
     }
 
+    /// Serialize transaction into consensus-valid format. See https://docs.rs/bitcoin/latest/bitcoin/struct.Transaction.html#serialization-notes for more notes on transaction serialization.
     pub fn serialize(&self) -> Vec<u8> {
         serialize(&self.0)
     }
 
+    /// List of transaction inputs.
     pub fn input(&self) -> Vec<TxIn> {
         self.0.input.iter().map(|tx_in| tx_in.into()).collect()
     }
 
+    /// List of transaction outputs.
     pub fn output(&self) -> Vec<TxOut> {
         self.0.output.iter().map(|tx_out| tx_out.into()).collect()
     }
 
+    /// Block height or timestamp. Transaction cannot be included in a block until this height/time.
+    ///
+    /// /// ### Relevant BIPs
+    ///
+    /// * [BIP-65 OP_CHECKLOCKTIMEVERIFY](https://github.com/bitcoin/bips/blob/master/bip-0065.mediawiki)
+    /// * [BIP-113 Median time-past as endpoint for lock-time calculations](https://github.com/bitcoin/bips/blob/master/bip-0113.mediawiki)
     pub fn lock_time(&self) -> u32 {
         self.0.lock_time.to_consensus_u32()
     }
@@ -322,25 +428,48 @@ impl From<&Transaction> for BdkTransaction {
     }
 }
 
+/// A Partially Signed Transaction.
+#[derive(uniffi::Object)]
 pub struct Psbt(pub(crate) Mutex<BdkPsbt>);
 
+#[uniffi::export]
 impl Psbt {
+    /// Creates a new `Psbt` instance from a base64-encoded string.
+    #[uniffi::constructor]
     pub(crate) fn new(psbt_base64: String) -> Result<Self, PsbtParseError> {
         let psbt: BdkPsbt = BdkPsbt::from_str(&psbt_base64)?;
         Ok(Psbt(Mutex::new(psbt)))
     }
 
+    /// Serialize the PSBT into a base64-encoded string.
     pub(crate) fn serialize(&self) -> String {
         let psbt = self.0.lock().unwrap().clone();
         psbt.to_string()
     }
 
+    /// Extracts the `Transaction` from a `Psbt` by filling in the available signature information.
+    ///
+    /// #### Errors
+    ///
+    /// `ExtractTxError` variants will contain either the `Psbt` itself or the `Transaction`
+    /// that was extracted. These can be extracted from the Errors in order to recover.
+    /// See the error documentation for info on the variants. In general, it covers large fees.
     pub(crate) fn extract_tx(&self) -> Result<Arc<Transaction>, ExtractTxError> {
         let tx: BdkTransaction = self.0.lock().unwrap().clone().extract_tx()?;
         let transaction: Transaction = tx.into();
         Ok(Arc::new(transaction))
     }
 
+    /// Calculates transaction fee.
+    ///
+    /// 'Fee' being the amount that will be paid for mining a transaction with the current inputs
+    /// and outputs i.e., the difference in value of the total inputs and the total outputs.
+    ///
+    /// #### Errors
+    ///
+    /// - `MissingUtxo` when UTXO information for any input is not present or is invalid.
+    /// - `NegativeFee` if calculated value is negative.
+    /// - `FeeOverflow` if an integer overflow occurs.
     pub(crate) fn fee(&self) -> Result<u64, PsbtError> {
         self.0
             .lock()
@@ -350,6 +479,9 @@ impl Psbt {
             .map_err(PsbtError::from)
     }
 
+    /// Combines this `Psbt` with `other` PSBT as described by BIP 174.
+    ///
+    /// In accordance with BIP 174 this function is commutative i.e., `A.combine(B) == B.combine(A)`
     pub(crate) fn combine(&self, other: Arc<Psbt>) -> Result<Arc<Psbt>, PsbtError> {
         let mut original_psbt = self.0.lock().unwrap().clone();
         let other_psbt = other.0.lock().unwrap().clone();
@@ -357,6 +489,9 @@ impl Psbt {
         Ok(Arc::new(Psbt(Mutex::new(original_psbt))))
     }
 
+    /// Finalizes the current PSBT and produces a result indicating
+    ///
+    /// whether the finalization was successful or not.
     pub(crate) fn finalize(&self) -> FinalizedPsbtResult {
         let curve = Secp256k1::verification_only();
         let finalized = self.0.lock().unwrap().clone().finalize(&curve);
@@ -377,6 +512,7 @@ impl Psbt {
         }
     }
 
+    /// Serializes the PSBT into a JSON string representation.
     pub(crate) fn json_serialize(&self) -> String {
         let psbt = self.0.lock().unwrap();
         serde_json::to_string(psbt.deref()).unwrap()
@@ -389,17 +525,23 @@ impl From<BdkPsbt> for Psbt {
     }
 }
 
+#[derive(uniffi::Record)]
 pub struct FinalizedPsbtResult {
     pub psbt: Arc<Psbt>,
     pub could_finalize: bool,
     pub errors: Option<Vec<PsbtFinalizeError>>,
 }
 
-#[derive(Debug, Clone)]
+/// A transcation input.
+#[derive(Debug, Clone, uniffi::Record)]
 pub struct TxIn {
+    /// A pointer to the previous output this input spends from.
     pub previous_output: OutPoint,
+    /// The script corresponding to the `scriptPubKey`, empty in SegWit transactions.
     pub script_sig: Arc<Script>,
+    /// https://bitcoin.stackexchange.com/questions/87372/what-does-the-sequence-in-a-transaction-input-mean
     pub sequence: u32,
+    /// A proof for the script that authorizes the spend of the output.
     pub witness: Vec<Vec<u8>>,
 }
 
@@ -417,9 +559,18 @@ impl From<&BdkTxIn> for TxIn {
     }
 }
 
-#[derive(Debug, Clone)]
+/// Bitcoin transaction output.
+///
+/// Defines new coins to be created as a result of the transaction,
+/// along with spending conditions ("script", aka "output script"),
+/// which an input spending it must satisfy.
+///
+/// An output that is not yet spent by an input is called Unspent Transaction Output ("UTXO").
+#[derive(Debug, Clone, uniffi::Record)]
 pub struct TxOut {
+    /// The value of the output, in satoshis.
     pub value: u64,
+    /// The script which must be satisfied for the output to be spent.
     pub script_pubkey: Arc<Script>,
 }
 
@@ -431,6 +582,42 @@ impl From<&BdkTxOut> for TxOut {
         }
     }
 }
+
+/// A bitcoin Block hash
+#[derive(Debug, Clone, Copy, PartialEq, Eq, std::hash::Hash, uniffi::Object)]
+#[uniffi::export(Display, Eq, Hash)]
+pub struct BlockHash(pub(crate) BitcoinBlockHash);
+
+impl_hash_like!(BlockHash, BitcoinBlockHash);
+
+/// A bitcoin transaction identifier
+#[derive(Debug, Clone, Copy, PartialEq, Eq, std::hash::Hash, uniffi::Object)]
+#[uniffi::export(Display, Eq, Hash)]
+pub struct Txid(pub(crate) BitcoinTxid);
+
+impl_hash_like!(Txid, BitcoinTxid);
+
+/// A bitcoin transaction identifier, including witness data.
+/// For transactions with no SegWit inputs, the `txid` will be equivalent to `wtxid`.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, std::hash::Hash, uniffi::Object)]
+#[uniffi::export(Display, Eq, Hash)]
+pub struct Wtxid(pub(crate) BitcoinWtxid);
+
+impl_hash_like!(Wtxid, BitcoinWtxid);
+
+/// A collision-proof unique identifier for a descriptor.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, std::hash::Hash, uniffi::Object)]
+#[uniffi::export(Display, Eq, Hash)]
+pub struct DescriptorId(pub(crate) BitcoinSha256Hash);
+
+impl_hash_like!(DescriptorId, BitcoinSha256Hash);
+
+/// The merkle root of the merkle tree corresponding to a block's transactions.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, std::hash::Hash, uniffi::Object)]
+#[uniffi::export(Display, Eq, Hash)]
+pub struct TxMerkleNode(pub(crate) BitcoinDoubleSha256Hash);
+
+impl_hash_like!(TxMerkleNode, BitcoinDoubleSha256Hash);
 
 #[cfg(test)]
 mod tests {