+++ bdk-ffi/src/descriptor.rs
@@ -19,13 +19,18 @@ use std::fmt::Display;
 use std::str::FromStr;
 use std::sync::Arc;
 
-#[derive(Debug)]
+/// An expression of how to derive output scripts: https://github.com/bitcoin/bitcoin/blob/master/doc/descriptors.md
+#[derive(Debug, uniffi::Object)]
+#[uniffi::export(Debug, Display)]
 pub struct Descriptor {
     pub extended_descriptor: ExtendedDescriptor,
     pub key_map: KeyMap,
 }
 
+#[uniffi::export]
 impl Descriptor {
+    /// Parse a string as a descriptor for the given network.
+    #[uniffi::constructor]
     pub fn new(descriptor: String, network: Network) -> Result<Self, DescriptorError> {
         let secp = Secp256k1::new();
         let (extended_descriptor, key_map) = descriptor.into_wallet_descriptor(&secp, network)?;
@@ -35,6 +40,8 @@ impl Descriptor {
         })
     }
 
+    /// Multi-account hierarchy descriptor: https://github.com/bitcoin/bips/blob/master/bip-0044.mediawiki
+    #[uniffi::constructor]
     pub fn new_bip44(
         secret_key: &DescriptorSecretKey,
         keychain_kind: KeychainKind,
@@ -61,6 +68,8 @@ impl Descriptor {
         }
     }
 
+    /// Multi-account hierarchy descriptor: https://github.com/bitcoin/bips/blob/master/bip-0044.mediawiki
+    #[uniffi::constructor]
     pub fn new_bip44_public(
         public_key: &DescriptorPublicKey,
         fingerprint: String,
@@ -92,6 +101,8 @@ impl Descriptor {
         }
     }
 
+    /// P2SH nested P2WSH descriptor: https://github.com/bitcoin/bips/blob/master/bip-0049.mediawiki
+    #[uniffi::constructor]
     pub fn new_bip49(
         secret_key: &DescriptorSecretKey,
         keychain_kind: KeychainKind,
@@ -118,6 +129,8 @@ impl Descriptor {
         }
     }
 
+    /// P2SH nested P2WSH descriptor: https://github.com/bitcoin/bips/blob/master/bip-0049.mediawiki
+    #[uniffi::constructor]
     pub fn new_bip49_public(
         public_key: &DescriptorPublicKey,
         fingerprint: String,
@@ -149,6 +162,8 @@ impl Descriptor {
         }
     }
 
+    /// Pay to witness PKH descriptor: https://github.com/bitcoin/bips/blob/master/bip-0084.mediawiki
+    #[uniffi::constructor]
     pub fn new_bip84(
         secret_key: &DescriptorSecretKey,
         keychain_kind: KeychainKind,
@@ -175,6 +190,8 @@ impl Descriptor {
         }
     }
 
+    /// Pay to witness PKH descriptor: https://github.com/bitcoin/bips/blob/master/bip-0084.mediawiki
+    #[uniffi::constructor]
     pub fn new_bip84_public(
         public_key: &DescriptorPublicKey,
         fingerprint: String,
@@ -206,6 +223,8 @@ impl Descriptor {
         }
     }
 
+    /// Single key P2TR descriptor: https://github.com/bitcoin/bips/blob/master/bip-0086.mediawiki
+    #[uniffi::constructor]
     pub fn new_bip86(
         secret_key: &DescriptorSecretKey,
         keychain_kind: KeychainKind,
@@ -232,6 +251,8 @@ impl Descriptor {
         }
     }
 
+    /// Single key P2TR descriptor: https://github.com/bitcoin/bips/blob/master/bip-0086.mediawiki
+    #[uniffi::constructor]
     pub fn new_bip86_public(
         public_key: &DescriptorPublicKey,
         fingerprint: String,
@@ -263,16 +284,19 @@ impl Descriptor {
         }
     }
 
+    /// Dangerously convert the descriptor to a string.
     pub fn to_string_with_secret(&self) -> String {
         let descriptor = &self.extended_descriptor;
         let key_map = &self.key_map;
         descriptor.to_string_with_secret(key_map)
     }
 
+    /// Does this descriptor contain paths: https://github.com/bitcoin/bips/blob/master/bip-0389.mediawiki
     pub fn is_multipath(&self) -> bool {
         self.extended_descriptor.is_multipath()
     }
 
+    /// Return descriptors for all valid paths.
     pub fn to_single_descriptors(&self) -> Result<Vec<Arc<Descriptor>>, MiniscriptError> {
         self.extended_descriptor
             .clone()
@@ -300,7 +324,8 @@ impl Display for Descriptor {
 
 #[cfg(test)]
 mod test {
-    use crate::*;
+    use super::*;
+    use crate::keys::{DerivationPath, Mnemonic};
     use assert_matches::assert_matches;
     use bdk_wallet::bitcoin::Network;
     use bdk_wallet::KeychainKind;
@@ -313,31 +338,31 @@ mod test {
     #[test]
     fn test_descriptor_templates() {
         let master: DescriptorSecretKey = get_descriptor_secret_key();
-        println!("Master: {:?}", master.as_string());
+        println!("Master: {:?}", master.to_string());
         // tprv8ZgxMBicQKsPdWuqM1t1CDRvQtQuBPyfL6GbhQwtxDKgUAVPbxmj71pRA8raTqLrec5LyTs5TqCxdABcZr77bt2KyWA5bizJHnC4g4ysm4h
         let handmade_public_44 = master
             .derive(&DerivationPath::new("m/44h/1h/0h".to_string()).unwrap())
             .unwrap()
             .as_public();
-        println!("Public 44: {}", handmade_public_44.as_string());
+        println!("Public 44: {}", handmade_public_44);
         // Public 44: [d1d04177/44'/1'/0']tpubDCoPjomfTqh1e7o1WgGpQtARWtkueXQAepTeNpWiitS3Sdv8RKJ1yvTrGHcwjDXp2SKyMrTEca4LoN7gEUiGCWboyWe2rz99Kf4jK4m2Zmx/*
         let handmade_public_49 = master
             .derive(&DerivationPath::new("m/49h/1h/0h".to_string()).unwrap())
             .unwrap()
             .as_public();
-        println!("Public 49: {}", handmade_public_49.as_string());
+        println!("Public 49: {}", handmade_public_49);
         // Public 49: [d1d04177/49'/1'/0']tpubDC65ZRvk1NDddHrVAUAZrUPJ772QXzooNYmPywYF9tMyNLYKf5wpKE7ZJvK9kvfG3FV7rCsHBNXy1LVKW95jrmC7c7z4hq7a27aD2sRrAhR/*
         let handmade_public_84 = master
             .derive(&DerivationPath::new("m/84h/1h/0h".to_string()).unwrap())
             .unwrap()
             .as_public();
-        println!("Public 84: {}", handmade_public_84.as_string());
+        println!("Public 84: {}", handmade_public_84);
         // Public 84: [d1d04177/84'/1'/0']tpubDDNxbq17egjFk2edjv8oLnzxk52zny9aAYNv9CMqTzA4mQDiQq818sEkNe9Gzmd4QU8558zftqbfoVBDQorG3E4Wq26tB2JeE4KUoahLkx6/*
         let handmade_public_86 = master
             .derive(&DerivationPath::new("m/86h/1h/0h".to_string()).unwrap())
             .unwrap()
             .as_public();
-        println!("Public 86: {}", handmade_public_86.as_string());
+        println!("Public 86: {}", handmade_public_86);
         // Public 86: [d1d04177/86'/1'/0']tpubDCJzjbcGbdEfXMWaL6QmgVmuSfXkrue7m2YNoacWwyc7a2XjXaKojRqNEbo41CFL3PyYmKdhwg2fkGpLX4SQCbQjCGxAkWHJTw9WEeenrJb/*
         let template_private_44 =
             Descriptor::new_bip44(&master, KeychainKind::External, Network::Testnet);
@@ -380,7 +405,7 @@ mod test {
         println!("Template public 44: {}", template_public_44);
         println!("Template public 84: {}", template_public_84);
         println!("Template public 86: {}", template_public_86);
-        // when using a public key, both to_string and as_string_private return the same string
+        // when using a public key, both to_string and to_string_private return the same string
         assert_eq!(
             template_public_44.to_string_with_secret(),
             template_public_44.to_string()