+++ bdk-ffi/tests/bindings/test.swift
@@ -8,6 +8,3 @@ import BitcoinDevKit
 
 // A type from the bitcoin-ffi library
 let network = Network.testnet
-
-// A type from the bdk-ffi library
-let blockId = BlockId(height: 32, hash: "abcd")